+++ tidy3d/plugins/mode/solver.py
@@ -24,9 +24,11 @@ TOL_TENSORIAL = 1e-6
 TARGET_SHIFT = 10 * fp_eps
 # Preconditioner: "Jacobi" or "Material" based
 PRECONDITIONER = "Material"
-# PEC permittivity cut-off value. Let it be as large as possible so long as not causing overflow in
+# Good conductor permittivity cut-off value. Let it be as large as possible so long as not causing overflow in
 # double precision. This value is very heuristic.
-PEC_CUT_OFF = 1e70
+GOOD_CONDUCTOR_CUT_OFF = 1e70
+# Consider a material to be good conductor if |ep| (or |mu|) > GOOD_CONDUCTOR_THRESHOLD * |pec_val|
+GOOD_CONDUCTOR_THRESHOLD = 0.9
 
 
 class EigSolver(Tidy3dBaseModel):
@@ -41,6 +43,7 @@ class EigSolver(Tidy3dBaseModel):
         coords,
         freq,
         mode_spec,
+        precision,
         mu_cross=None,
         split_curl_scaling=None,
         symmetry=(0, 0),
@@ -64,6 +67,8 @@ class EigSolver(Tidy3dBaseModel):
             (Hertz) Frequency at which the eigenmodes are computed.
         mode_spec : ModeSpec
             ``ModeSpec`` object containing specifications of the mode solver.
+        precision : Literal["single", "double"]
+            Apply "single" or "double" precision in mode solver.
         mu_cross : array_like or tuple of array_like
             Either a single 2D array defining the relative permeability in the cross-section,
             or nine 2D arrays defining the permeability at the Hx, Hy, and Hz locations
@@ -249,7 +254,7 @@ class EigSolver(Tidy3dBaseModel):
             der_mats,
             num_modes,
             target_neff_p,
-            mode_spec.precision,
+            precision,
             direction,
             enable_incidence_matrices,
             basis_E=basis_E,
@@ -330,20 +335,22 @@ class EigSolver(Tidy3dBaseModel):
 
         # In the matrices P and Q below, they contain terms ``epsilon_parallel`` or
         # ``mu_parallel``, and also a term proportional to 1/(k0 * dl)**2. To make sure
-        # that permittivity of PEC is visible in low-frequency/high resolution, pec_val should be
+        # that permittivity of good conductor is visible in low-frequency/high resolution, pec_val should be
         # scaled by a factor max(1, max(1/k0 dl) **2).
         pec_scaling = max(1, max([np.max(abs(f)) for f in der_mats]) ** 2)
-        pec_scaled_val = min(PEC_CUT_OFF, pec_scaling * abs(pec_val))
+        pec_scaled_val = min(GOOD_CONDUCTOR_CUT_OFF, pec_scaling * abs(pec_val))
 
-        # use a high-conductivity model for locations associated with a PEC
-        def conductivity_model_for_pec(eps, threshold=0.9 * pec_val):
-            """PEC entries associated with 'eps' are converted to a high-conductivity model."""
+        # use a high-conductivity model for locations associated with a good conductor
+        def conductivity_model_for_good_conductor(
+            eps, threshold=GOOD_CONDUCTOR_THRESHOLD * pec_val
+        ):
+            """Entries associated with 'eps' are converted to a high-conductivity model."""
             eps = eps.astype(complex)
             eps[np.abs(eps) >= abs(threshold)] = 1 + 1j * pec_scaled_val
             return eps
 
-        eps_tensor = conductivity_model_for_pec(eps_tensor)
-        mu_tensor = conductivity_model_for_pec(mu_tensor)
+        eps_tensor = conductivity_model_for_good_conductor(eps_tensor)
+        mu_tensor = conductivity_model_for_good_conductor(mu_tensor)
 
         # Determine if ``eps`` and ``mu`` are diagonal or tensorial
         off_diagonals = (np.ones((3, 3)) - np.eye(3)).astype(bool)
@@ -473,11 +480,10 @@ class EigSolver(Tidy3dBaseModel):
         mu_zz = mu[2, 2, :]
         dxf, dxb, dyf, dyb = der_mats
 
-        def any_pec(eps_vec, threshold=0.9 * np.abs(pec_val)):
-            """Check if there are any PEC values in the given permittivity array."""
-            return np.any(np.abs(eps_vec) >= threshold)
-
-        if any(any_pec(i) for i in [eps_xx, eps_yy, eps_zz, mu_xx, mu_yy, mu_zz]):
+        if any(
+            cls.mode_plane_contain_good_conductor(i)
+            for i in [eps_xx, eps_yy, eps_zz, mu_xx, mu_yy, mu_zz]
+        ):
             enable_preconditioner = True
 
         # Compute the matrix for diagonalization
@@ -1010,6 +1016,15 @@ class EigSolver(Tidy3dBaseModel):
     def split_curl_field_postprocess_inverse(split_curl, E):
         """E has the shape (3, N, num_modes)"""
         raise RuntimeError("Split curl not yet implemented for relative mode solver.")
+
+    @staticmethod
+    def mode_plane_contain_good_conductor(material_response) -> bool:
+        """Find out if epsilon on the modal plane contain good conductors whose permittivity
+        or permeability value is very large.
+        """
+        if material_response is None:
+            return False
+        return np.any(np.abs(material_response) > GOOD_CONDUCTOR_THRESHOLD * np.abs(pec_val))
 
 
 def compute_modes(*args, **kwargs) -> Tuple[Numpy, Numpy, str]: